feat: Lookup all tags by date first (#438)
* feat: Lookup all tags by date first

In repositories with multiple packages, a new tag may not be created for every project at the same time.
By looking up all the tags at once, we can ensure that we get the most recent tag for each package, not
just the most recent tag for the entire repository.

* chore: Update tests