fix(plugin): Walk all subtrees (#716)
Due to short-circuiting, the `walkAST` function was never called on subsequent children once the first matching subtree was found. From MDN: "The logical OR assignment (||=) operator only evaluates the right operand and assigns to the left if the left operand is falsy.". By walking the child's subtree first, we will walk all possible trees.