chore: Remove helper export (#741)
If the helper export from the root index is used in the consuming app's
code, Vite treats the entire module as code to be executed in the
browser. This causes build errors. Removing the export from the root
index forces users to use the full path to the helper, which _is_ valid
browser code.